现场调试之后的最终版本，分为 3台和2台的区别。

+++ pressuresetdialog.h
@@ -17,7 +17,7 @@ public:
     ~PressureSetDialog();
 
 signals:
-    void closeCurrentDialog(int maxPressure,int minPressure);
+    void closeCurrentDialog(float maxPressure,float minPressure);
 
 private slots:
     void on_setPressureBtn_clicked();
